折れ線グラフのコンポーネント化
